requirements.txt
s3fs==2022.5.0
boto3==1.21.21
numpy<1.24.0
pandas==1.3.5
tqdm==4.64.0
openpyxl
plac
faiss-cpu
pyyaml
transformers
sentence-transformers
toolz